# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repo-finisher/status:
    get:
      operationId: getRepoFinisherStatus
      tags: [repo-finisher]
      summary: Check whether a repo has previously been finished
      security:
        - bearerAuth: []
      parameters:
        - name: repo
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Finish status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoFinisherStatus"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    FinishChange:
      type: object
      required: [file, status, description]
      properties:
        file:
          type: string
        status:
          type: string
          enum: [created, modified, deleted]
        description:
          type: string
    SkippedChange:
      type: object
      required: [path, reason]
      properties:
        path:
          type: string
        reason:
          type: string
    ExecuteResult:
      type: object
      required:
        [repo, planId, branch, commit_sha, pr_url, pr_number, files_changed, summary, changes, skipped]
      properties:
        repo:
          type: string
        planId:
          type: string
        branch:
          type: string
        commit_sha:
          type: string
        pr_url:
          type: string
        pr_number:
          type: integer
        files_changed:
          type: integer
        summary:
          type: string
        changes:
          type: array
          items:
            $ref: "#/components/schemas/FinishChange"
        skipped:
          type: array
          items:
            $ref: "#/components/schemas/SkippedChange"
    RepoFinisherStatus:
      type: object
      required: [repo, hasBeenFinished, finishes]
      properties:
        repo:
          type: string
        hasBeenFinished:
          type: boolean
        finishes:
          type: array
          items:
            $ref: "#/components/schemas/ExecuteResult"
...